backend/requirements.txt
# CivicLens AI - Backend Dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.13.0
aiosqlite>=0.19.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
httpx>=0.25.2
beautifulsoup4>=4.12.2
lxml>=5.0.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
jinja2>=3.1.2
aiofiles>=23.2.1
python-dateutil>=2.8.2

# Feature 1: Document Validation
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Feature 2: Push Notifications
pywebpush>=2.0.0
